Modifications to mt GD (working)
Error is still pretty bad but it is very close to sklearn's, so I'm
mostly convinced this is it for a linear regressor with bare features.

Also, there is a min-max scaler that fits to train data and is applied
to test data afterwards.

Training vs test error plot is garbage for this model.