Added CNN model for feature extraction
So, I've been reading this https://bitbucket.org/polimi-ispl/
camera-model-identification-with-cnn and I thought: why don't
I use the feature-extraction layers from this guy and work the
classification myself?

Yeah, nice, but I think I might stumble upon some problems (as
I already did to reproduce the model in Keras) so I'm putting
thin in the fridge for a while.